Fix issues from code review
