+++ frontend/src/__test__/features/users/stores/reducers/operator/authOperation.spec.ts
@@ -5,10 +5,10 @@ import {
   pendingOperation,
   fulfilledOperation,
   rejectedOperation,
-} from "@/features/users/stores/reducers/authReducers";
+} from "@/features/users/stores/reducers/operator/authOperation";
 import type { AuthState, AuthResponse } from "@/features/users/types/authTypes";
 
-describe("【ユニットテスト】State操作に関わるヘルパー関数(ユーザー登録)のテスト。", () => {
+describe("【ユニットテスト】State操作に関わる操作関数(ユーザー登録)のテスト。", () => {
   let state: AuthState;
   beforeEach(() => {
     state = {
